componentize deploy components
