+++ components/getURLparameterValue.tsx
@@ -1,61 +1,4 @@
-//
-// sign up with agora, on the sso check if the redirection happening.
-//
-//
-
-// import { SettingsApplications } from "@material-ui/icons";
-// import { profile } from "console";
-// import { argsToArgsConfig } from "graphql/type/definition";
-// import app from "next/app";
-
-// export default function (url_string: string) {
-//   const url = new URL(url_string);
-//   return url.searchParams;
-// }
-
-// Attendee leave callback on working
-//
-// In all react-native mobile application, screen sharing binking.
-
-// screen sharing blinking seen so far only on app builder app, test on a test test-react-native app to verify.
-// note: the sdk is throwing mute and unmute rapidly when someone is sscreen share on react-native.
-// 5 times every seconds.
-
-// rtc profile
-// communication profile
-
-// just a normal react-native app, install agora
-// screen sharing blinking, quick start app from ekaansh
-
-// user-left-event occurs, we are using to maintain the users state in UIKIt
-
-// Two profiles type in Agora while creating the client object using the key called 'mode'
-
-//  - rtc profile
-//  - communication(live) profile
-
-//  In Live-streaming we'll change the mode from rtc to live.
-//  User leave event works only for the host users
-//  User leave callback is not triggered for attendee.
-//  This POC most validate the user-leave callback for web and react-native.
-//  How to make the leave-callback trigger for the attendees.
-
-//
-//  In Live-streaming we'll change the mode from rtc to live.
-//  the user leave event works only for the hosts (17)
-//  how to make it work for the attendees.
-
-//  difference: attendee and host separation, rtc gives you the methods for these separation.
-//
-
-//
-//  they give you a role. when sdk says role as attendee they can not publish (), they can only subscribe.
-//
-//
-//
-// subscribe to agora channel, audience can not publish.
-//
-//
-//
-//
-//
+export default function (url_string: string) {
+  const url = new URL(url_string);
+  return url.searchParams;
+}
