+++ components/icons.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+const close = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M4.49512 19.5045L19.5045 4.49512"
+        stroke="#5E6871"
+        strokeWidth="1.5"
+      />
+      <path
+        d="M4.49512 4.49512L19.5045 19.5045"
+        stroke="#5E6871"
+        strokeWidth="1.5"
+      />
+    </svg>
+  );
+};
+
+export {close};
